Centralized transactional data into TransactionSummary.

+++ app/src/main/java/com/example/budgetassistant/models/BankAccount.java
@@ -2,11 +2,11 @@ package com.example.budgetassistant.models;
 
 public class BankAccount {
     public String AccountNumber;
-    public int Balance;
+    public float Balance;
     public String BankName;
     public String DisplayName;
 
-    public BankAccount(String accountNumber, int balance, String bankName, String displayName) {
+    public BankAccount(String accountNumber, float balance, String bankName, String displayName) {
         AccountNumber = accountNumber;
         Balance = balance;
         BankName = bankName;
